+++ src/blacklist.rs
@@ -28,9 +28,12 @@ impl MutationChecksum {
         Self(format!("{:x}", md5::compute(lines)))
     }
 
-    #[cfg(test)]
     pub(crate) fn as_str(&self) -> &str {
         &self.0
+    }
+
+    pub(crate) fn into_string(self) -> String {
+        self.0
     }
 
     fn parse(value: &str) -> Option<Self> {